# swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "GP Premium API",
        "description": "API para gerenciamento de produção de pneus recauchutados da GP Premium",
        "contact": {
            "name": "Equipe GP Premium",
            "url": "https://www.gppremium.com.br",
            "email": "contato@gppremium.com.br"
        },
        "license": {
            "name": "Proprietary",
            "url": "https://www.gppremium.com.br/license"
        },
        "version": "0.0.35-SNAPSHOT"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Servidor de Desenvolvimento"
        },
        {
            "url": "https://api.gppremium.com.br",
            "description": "Servidor de Produção"
        }
    ],
    "security": [
        {
            "Bearer Authentication": []
        }
    ],
...
    "paths": {
...
        "/api/tipoobservacao/{id}": {
            "get": {
                "tags": [
                    "tipo-observacao-controller"
                ],
                "operationId": "consultar_3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "tipo-observacao-controller"
                ],
                "operationId": "atualizar",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TipoObservacaoModel"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "tipo-observacao-controller"
                ],
                "operationId": "delete_1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/tipoclassificacao/{id}": {
            "get": {
                "tags": [
                    "tipo-classificacao-controller"
                ],
                "operationId": "consultar_4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "tipo-classificacao-controller"
                ],
                "operationId": "atualizar_1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TipoClassificacaoModel"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "tipo-classificacao-controller"
                ],
                "operationId": "delete_2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/regra/{id}": {
            "get": {
                "tags": [
                    "regra-controller"
                ],
                "operationId": "consultar_5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "regra-controller"
                ],
                "operationId": "atualizar_2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegraModel"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "regra-controller"
                ],
                "operationId": "delete_3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "TipoClassificacaoModel": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "descricao": {
                        "type": "string"
                    }
                }
            },
            "TipoObservacaoModel": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "descricao": {
                        "type": "string"
                    },
                    "tipo_classificacao": {
                        "$ref": "#/components/schemas/TipoClassificacaoModel"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "Bearer Authentication": {
                "type": "http",
                "description": "Token JWT para autenticação. Formato: Bearer {token}",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}